wire profile into the pipeline
